requirements.txt
flask
torch
torchvision
faiss-cpu
Pillow
